@@ -6,6 +6,6 @@ public partial class App : Application
 	{
 		InitializeComponent();
 
-		MainPage = new NavigationPage(new MainPage());
+		MainPage = new AppFlyoutPage { Detail = new NavigationPage(new MainPage1()) };
 	}
 }
